+++ packages/cli/src/operations.ts
@@ -9,24 +9,7 @@ import { get } from 'lodash';
 import type { Spec } from 'swagger-schema-official';
 import type { OpenAPIObject } from 'openapi3-ts';
 import type { CollectionDefinition } from 'postman-collection';
-import { isSome } from 'fp-ts/lib/Option';
-import {
-  stripReadOnlyProperties,
-  stripWriteOnlyProperties,
-  IHttpOperationEx,
-  JSONSchema,
-  JSONSchemaEx,
-  createJsonSchemaFromParams,
-} from '@stoplight/prism-http';
-
-const bundle = (schema: JSONSchema, bundle?: unknown): JSONSchemaEx => {
-  return {
-    ...schema,
-    __bundled__: bundle,
-  };
-};
-
-const EMPTY_SCHEMA: JSONSchema = {};
+import { IHttpOperationEx, enrichWithPreGeneratedValidationSchema } from '@stoplight/prism-http';
 
 export async function getHttpOperationsFromSpec(specFilePathOrObject: string | object): Promise<IHttpOperationEx[]> {
   const result = decycle(await dereference(specFilePathOrObject));
@@ -37,63 +20,13 @@ export async function getHttpOperationsFromSpec(specFilePathOrObject: string | o
   else if (isPostmanCollection(result)) operations = transformPostmanCollectionOperations(result);
   else throw new Error('Unsupported document format');
 
+  enrichWithPreGeneratedValidationSchema(operations);
+
   operations.forEach((op, i, ops) => {
-    const opEx = op as IHttpOperationEx;
-    if (opEx.request !== undefined) {
-      opEx.request.headersValidatingSchema = EMPTY_SCHEMA;
-      if (opEx.request.headers !== undefined) {
-        opEx.request.headersValidatingSchema = bundle(
-          createJsonSchemaFromParams(opEx.request.headers),
-          opEx['__bundled__']
-        );
-      }
-
-      opEx.request.pathValidatingSchema = EMPTY_SCHEMA;
-      if (opEx.request.path !== undefined) {
-        opEx.request.pathValidatingSchema = bundle(createJsonSchemaFromParams(opEx.request.path), opEx['__bundled__']);
-      }
-
-      opEx.request.queryValidatingSchema = EMPTY_SCHEMA;
-      if (opEx.request.query !== undefined) {
-        opEx.request.queryValidatingSchema = bundle(
-          createJsonSchemaFromParams(opEx.request.query),
-          opEx['__bundled__']
-        );
-      }
-
-      if (opEx.request.body !== undefined) {
-        opEx.request.body.contents?.forEach(mtc => {
-          mtc.contentValidatingSchema = EMPTY_SCHEMA;
-          if (mtc.schema !== undefined) {
-            const newLocal = stripReadOnlyProperties(mtc.schema);
-            if (isSome(newLocal)) {
-              mtc.contentValidatingSchema = bundle(newLocal.value, opEx['__bundled__']);
-            }
-          }
-        });
-      }
-
-      opEx.responses.forEach(res => {
-        res.headersValidatingSchema = EMPTY_SCHEMA;
-        if (res.headers !== undefined) {
-          res.headersValidatingSchema = bundle(createJsonSchemaFromParams(res.headers), opEx['__bundled__']);
-        }
-
-        res.contents?.forEach(mtc => {
-          mtc.contentValidatingSchema = EMPTY_SCHEMA;
-          if (mtc.schema !== undefined) {
-            const newLocal = stripWriteOnlyProperties(mtc.schema);
-            if (isSome(newLocal)) {
-              mtc.contentValidatingSchema = bundle(newLocal.value, opEx['__bundled__']);
-            }
-          }
-        });
-      });
-    }
     ops[i] = bundleTarget({
       document: {
         ...result,
-        __target__: opEx,
+        __target__: op,
       },
       path: '#/__target__',
       cloneDocument: false,
