+++ packages/http/src/index.ts
@@ -12,6 +12,7 @@ export * from './mocker/serializer/style';
 export { generate as generateHttpParam } from './mocker/generator/HttpParamGenerator';
 export { stripReadOnlyProperties, stripWriteOnlyProperties } from './utils/filterRequiredProperties';
 export { createJsonSchemaFromParams } from './validator/validators/params';
+export { enrichWithPreGeneratedValidationSchema } from './operations';
 import { IHttpConfig, IHttpResponse, IHttpRequest, PickRequired, PrismHttpComponents, IHttpProxyConfig } from './types';
 
 export const createInstance = (
